requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via django
camelot-py[cv]==0.9.0
    # via pdfwordify
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
chardet==5.2.0
    # via camelot-py
charset-normalizer==3.3.2
    # via
    #   pdfminer-six
    #   requests
click==8.1.7
    # via camelot-py
cryptography==42.0.5
    # via pdfminer-six
django==5.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.15.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
idna==3.7
    # via requests
lxml==5.2.1
    # via python-docx
multipledispatch==1.0.0
    # via pdfwordify
numpy==1.26.4
    # via
    #   camelot-py
    #   opencv-python
    #   pandas
opencv-python==4.9.0.80
    # via camelot-py
openpyxl==3.1.2
    # via camelot-py
packaging==24.0
    # via pytesseract
pandas==2.2.2
    # via camelot-py
pdfminer-six==20231228
    # via
    #   camelot-py
    #   pdfwordify
pdfwordify==0.0.1
    # via -r requirements.in
pillow==10.3.0
    # via
    #   pdfwordify
    #   pytesseract
pycparser==2.22
    # via cffi
pymupdf==1.24.2
    # via pdfwordify
pymupdfb==1.24.1
    # via pymupdf
pypdf2==2.12.1
    # via
    #   camelot-py
    #   pdfwordify
pytesseract==0.3.10
    # via pdfwordify
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.1.0
    # via pdfwordify
pytz==2024.1
    # via pandas
requests==2.31.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.5.0
    # via django
typing-extensions==4.11.0
    # via python-docx
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
